backend/requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
pocketbase==0.15.0
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
requests
websockets
runpod
fal-client
python-dotenv
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
